made into int

+++ my_tests/test_torque.py
@@ -62,7 +62,7 @@ world.scene.add(robot)
 
 
 # Define joint indices for the wheels (you can double-check them with the USD file)
-joint_indices = np.array([robot.get_dof_index(name) for name in wheel_dof_names])  # Should return correct indices
+joint_indices = np.array([robot.get_dof_index(name) for name in wheel_dof_names], dtype=np.int32)  # Should return correct indices
 
 def apply_wheel_torques(articulation_view, torques, indices):
     articulation_view.set_joint_efforts(efforts=torques, joint_indices=indices)
